DOCKER-257: fix 'docker inspect/history ...' for non-head image ids
`imgFromName` is the new canonical way to get from a repo (and tag)
name string (typically direct from a CLI command) to an image for
that user. Introduce AmbiguousDockerIdError to handle the current
design implication that the same docker id imported from separate
registries cannot be unique identified with the docker id.