DOCKER-458: Can't specify memory for containers with 1.7 client
Instead of getting the value Memory from just the root of the container
object, which is where docker clients < 1.7 put it, we also check in
container.HostConfig.Memory, which is where docker clients >= 1.7 put
it.

This commit adds integration tests for this change. These tests create
docker containers with a memory value different than the default memory
value and mimic the behavior of docker clients pre and post 1.7.

This commit also moves loadConfigSync to its own module so that it can
be used by tests that need to load the configuration to lookup default
values (such as defaultMemory in this case).

PR: #43
PR-URL: https://github.com/joyent/sdc-docker/pull/43
Reviewed-By: Josh Wilsdon <jwilsdon@joyent.com>